streamlined the build process
We now can submit the same .changes/.dsc/.tgz files to the autobuilder
for Fremantle and Diablo.
I changed debian/optify to 'auto', so now the auto-builder will
automatically 'optify' the package in fremantle, and leave it alone
in diablo.
I also commented out the py_compilefiles part (for now), since the diablo
auto-builder doesn't like it.