Hopefully fix bug #5672
This will hopefully fix bug #5672 "meveon crashes when started without a
network connection". We fix this by using python-conic to connect to the
default network when mevemon stops, or if there is no default netowork,
pop up the system dialog for selecting a connection.

I also changed the "except eveapi.Error:" sections to just "except:",
probably not a great idea, but we were crashing on socket exceptions
sometimes too (see bug #5663).