Added live 'total sp' counter in fremantle
We now display the total sp (actual sp + sp gained from
skill-in-training). This has only been enabled in fremantle, so diablo
gui will need to be adapted to take advantage of it.

The total sp is also 'live', meaning that it increments every second
based on the calculated SP/hour.