Stop the SP counter when window not on top
Previously our timer that increments the 'total sp' label kept ticking
away even when the window for the character sheet wasn't
visibile. The timer callback now checks if the window is visible (with
get_is_topmost()), and if not, stops the timer.

The Diablo GUI may need to try something different, not sure.