Convert API key to uppercase on input
A user was having problems getting his credentials to work, it turns out that
the API key is case sensitive.

We now automatically convert the letters in the API key to
upper-case before adding it to gconf. This was the user can input either
lower-case or upper-case letters, and both will work.